Moves Swagger setup to routes file
Moves the Swagger setup function from `router.go` to `routes.go`.
This change centralizes route configuration and improves code organization.
Also, updates the file path for the OpenAPI specification in the bump workflow.
